fix: resolve 108 bugs across all modules (Cat A/B/C)
Category A (8 critical): singleton thread-safety, blocking I/O on server
thread, SQL injection prevention, ExecutionSystem double-init guard.

Category B (42 important): thread-safe collections (ConcurrentHashMap,
AtomicInteger, volatile+DCL singletons), multiplayer correctness (BossBar
cleanup on disconnect, per-player quest state, HOLD_ZONE multi-player),
performance (ArrayDeque FIFO, persistent BufferedWriter, O(1) name index,
ellipse scan optimization), async file I/O, monotonic clocks, stamina
consumptionMultiplier application, 64-bit content hashing.

Category C (58 design): zero-vector dash crash, double-decrement tick bug,
wrong combo ActionType, EnumSet.copyOf empty set crashes, ZoneBounds
int overflow, HologramRegistry Optional return, TemplateRegistry double-
revision, TOCTOU cooldown fix, reflection->MethodHandle, SkinFetcher race
condition, PortalLinkerItem NPE, clone behavior mode clamp, 57k block scan
caching, SQL parameterization, ticket comment count sync, GLFW key constants,
dead code removal, magic number extraction, code deduplication (transport
registry 64->5 lines, layout helpers, BossBar color, slot list building).

Also adds TesterModality config infrastructure (Config.TESTER_MODALITY,
ChannelId 250-251, sync/ack payloads, compatibility test).

Build: 0 errors, 3957 tests pass, 0 failures.

+++ src/main/java/com/devmod/zone/data/ZoneBounds.java
@@ -174,9 +174,11 @@ public record ZoneBounds(
 
     /**
      * Gets the volume in blocks.
+     * Uses long arithmetic to prevent overflow for large zones
+     * (max zone size 1000^3 exceeds Integer.MAX_VALUE).
      */
-    public int volume() {
-        return width() * height() * length();
+    public long volume() {
+        return (long) width() * (long) height() * (long) length();
     }
 
     /**
